Remove restriction in custom_res against top-level directory
Currently, custom_res rule enforces that dir_name is not in the same
directory as the package. It does this implicitly:

```
old_resource_path.replace("/" + dir_name, "/res")
```

This means that if `"/" + dir_name` does not exist in the resource path,
nothing will be replaced, and the rule fails. Meaning that, for instance:

```
.
├── BUILD.bazel
├── res
├── res-debug
└── res-release
```

This fails, because the resource path is `res-debug/values/example.xml`.

By the nature of the rule, the dir_name ending in a slash feels more correct
to enforce than starting with a slash.